Add WebSocket functionality and update dependencies
This commit introduces a WebSocket functionality to the project by adding a new "socket.go" file in "internal/websocket/" directory. The WebSocket will allow the application to maintain a two-way communication channel over a single TCP connection. This also ensures a faster and real-time communication. It contains functions such as Socket(), Broadcast() and Clients(), which will create a new Handler, to broadcast message to all clients and to get all the clients respectively.

The "go.sum" file has also been updated to reflect changes in the dependencies likely due to the newly added file/feature.